Back detach ingredient app (#108)
* Ingredient app added

* urls added to ingredients.py

* views added to ingredients.py

* serializers added to ingredients app

* cleaning code

* urls, views, serializers moved from api to recipe app

* measurement & ingredient urls included into prostoemenu.urls

* Save settings development.py

* fix imports in api/urls.py

* fix urls for ingredient app

* fix urls for recipe app

* clean code

* * account, auth, kaiten urls moved into api.urls;
* recipe.urls, ingredient.urls path duplication fixed.

* Delete back/ingredient/admin.py

Ошибочное добавление файла

* * fix indents in prostoemenu/urls.py
