+++ back/prostoemenu/settings/development.py
@@ -16,14 +16,14 @@ DATABASES = {
     }
 
     # не удалять! Я локально работаю с посгресом.
-    #     'default': {
-    #         'ENGINE': os.getenv(
-    #             'DB_ENGINE', default='django.db.backends.postgresql'
-    #         ),
-    #         'NAME': os.getenv('DB_NAME', default='prostoemenu'),
-    #         'USER': os.getenv('POSTGRES_USER_DEV', default='postgres'),
-    #         'PASSWORD': os.getenv('POSTGRES_PASSWORD_DEV', default='postgres'),
-    #         'HOST': os.getenv('DB_HOST_DEV', default='db'),
-    #         'PORT': os.getenv('DB_PORT', default='5432')
-    #     }
+    # 'default': {
+    #     'ENGINE': os.getenv(
+    #         'DB_ENGINE', default='django.db.backends.postgresql'
+    #     ),
+    #     'NAME': os.getenv('DB_NAME', default='prostoemenu'),
+    #     'USER': os.getenv('POSTGRES_USER_DEV', default='postgres'),
+    #     'PASSWORD': os.getenv('POSTGRES_PASSWORD_DEV', default='postgres'),
+    #     'HOST': os.getenv('DB_HOST_DEV', default='db'),
+    #     'PORT': os.getenv('DB_PORT', default='5432')
+    # }
 }
