+++ back/prostoemenu/urls.py
@@ -1,23 +1,22 @@
-from django.contrib import admin
-from django.urls import path, include
 from django.conf import settings
 from django.conf.urls.static import static
-from drf_spectacular.views import SpectacularAPIView, SpectacularRedocView, SpectacularSwaggerView
+from django.contrib import admin
 
+from django.urls import (path,
+                         include)
+from drf_spectacular.views import (SpectacularAPIView,
+                                   SpectacularRedocView,
+                                   SpectacularSwaggerView)
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include('recipe.urls')),
+    path('api/', include('api.urls')),
     path('api/schema/', SpectacularAPIView.as_view(), name='schema'),
     path('api/schema/swagger/', SpectacularSwaggerView.as_view(url_name='schema'), name='swagger'),
     path('api/schema/redoc/', SpectacularRedocView.as_view(url_name='schema'), name='redoc'),
-    path('api/', include('api.urls')),
-    path('api/', include('kaiten.urls')),
-    path('account/', include('account.urls')),
-    path('auth/', include('djoser.urls')),
-
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
 
 if settings.DEBUG:
     import debug_toolbar
+
     urlpatterns += path('__debug__/', include('debug_toolbar.urls')),
