+++ back/recipe/serializers.py
@@ -1,4 +1,3 @@
-from rest_framework import serializers
 from recipe.models import (Ingredient,
                            Step,
                            Photo,
@@ -7,19 +6,10 @@ from recipe.models import (Ingredient,
                            RecipeSteps,
                            RecipeIngredients)
 from django.db import IntegrityError
+
+from ingredient.serializers import IngredientSerializer
+from rest_framework import serializers
 from rest_framework.exceptions import APIException
-
-
-class IngredientSerializerAllFields(serializers.ModelSerializer):
-    class Meta:
-        model = Ingredient
-        fields = '__all__'
-
-
-class IngredientSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Ingredient
-        fields = ('id', 'name', 'sort')
 
 
 class RecipeIngredientAlternative(serializers.ModelSerializer):
